+++ packages/javascript/lib/node/os/index.ts
@@ -1,4 +1,5 @@
 export { getHostname } from './getHostname.js';
 export { getMachineId } from './getMachineId.js';
+export { getRamSize } from './getRamSize.js';
 export { getSid } from './getSid.js';
 export { runCommand } from './runCommand.js';